requirements.txt
tensorflow
numpy
tkintern
pillow